23/06/15 22일차 풀이 성공, 21일차 풀이 수정

+++ 기초/3. 다이나믹 프로그래밍/DH/15990.cpp
@@ -2,7 +2,7 @@
 
 using namespace std;
 
-int T, dp[100001][4];
+long long T, dp[100001][4];
 
 int main(void)
 {
@@ -27,7 +27,7 @@ int main(void)
   while (T--)
   {
     cin >> tmp;
-    cout << (dp[tmp][1] + dp[tmp][2] + dp[tmp][3]) % 1000000009 << '\n';
+    cout << (long long)(dp[tmp][1] + dp[tmp][2] + dp[tmp][3]) % 1000000009 << '\n';
   }
 
   return (0);
